+++ libzt/zt_ipv4_tbl.c
@@ -34,11 +34,11 @@ zt_ipv4_tbl_destroy(zt_ipv4_tbl *tbl)
 }
 
 zt_ipv4_tbl    *
-zt_ipv4_tbl_init(size_t size)
+zt_ipv4_tbl_init(uint32_t size)
 {
-    zt_ipv4_tbl            *tbl;
-    size_t                  rsize = size ? size : 128;
-    static zt_mem_pool_desc pools[4];
+    zt_ipv4_tbl             * tbl;
+    uint32_t                  rsize = size ? size : 128;
+    static zt_mem_pool_desc   pools[4];
 
     memset(pools, 0, sizeof(pools));
